No need to delete termmeta if using the new tables

+++ includes/admin/class-wc-admin-taxonomies.php
@@ -79,7 +79,7 @@ class WC_Admin_Taxonomies {
 
 		$term_id = absint( $term_id );
 
-		if ( $term_id ) {
+		if ( $term_id && get_option( 'db_version' ) < 34370 ) {
 			$wpdb->delete( $wpdb->woocommerce_termmeta, array( 'woocommerce_term_id' => $term_id ), array( '%d' ) );
 		}
 	}
